// comcora-web-app-main/src/types/local.types.ts
export type AccountCurrencyCode =
  | "EUR"
  | "USD"
  | "CHF"
  | "DKK"
  | "SEK"
  | "RUB"
  | "GBP"
  | "JPY"
  | "INR"
  | "KRW"
  | "AUD"
  | "CAD"
  | "CNY"
  | "NZD"
  | "SGD"
  | "HKD"
  | "ZAR"
  | "BRL"
  | "TRY"
  | "MXN"
  | "NOK"
  | "PLN";

export type AccountCurrencySymbol = "€" | "$" | "£";